+++ scripts/check_version_in_changelog.py
@@ -13,7 +13,7 @@ if __name__ == '__main__':
     if not CHANGELOG_PATH.is_file():
         raise RuntimeError('Unable to find CHANGELOG.md file')
 
-    with open(CHANGELOG_PATH) as changelog_file:
+    with open(CHANGELOG_PATH, encoding='utf-8') as changelog_file:
         for line in changelog_file:
             # The heading for the changelog entry for the given version can start with either the version number, or the version number in a link
             if re.match(fr'\[?{current_package_version}([\] ]|$)', line):