+++ scripts/utils.py
@@ -2,15 +2,15 @@ import pathlib
 
 PACKAGE_NAME = 'apify'
 REPO_ROOT = pathlib.Path(__file__).parent.resolve() / '..'
-VERSION_FILE_PATH = REPO_ROOT / f'src/{PACKAGE_NAME}/_version.py'
+PYPROJECT_TOML_FILE_PATH = REPO_ROOT / 'pyproject.toml'
 
 
-# Load the current version number from src/package_name/_version.py
-# It is on a line in the format __version__ = 1.2.3
+# Load the current version number from pyproject.toml
+# It is on a line in the format `version = "1.2.3"`
 def get_current_package_version() -> str:
-    with open(VERSION_FILE_PATH, 'r') as version_file:
-        for line in version_file:
-            if line.startswith('__version__'):
+    with open(PYPROJECT_TOML_FILE_PATH, 'r', encoding='utf-8') as pyproject_toml_file:
+        for line in pyproject_toml_file:
+            if line.startswith('version = '):
                 delim = '"' if '"' in line else "'"
                 version = line.split(delim)[1]
                 return version
@@ -18,21 +18,21 @@ def get_current_package_version() -> str:
             raise RuntimeError('Unable to find version string.')
 
 
-# Write the given version number from src/package_name/_version.py
-# It replaces the version number on the line with the format __version__ = 1.2.3
+# Write the given version number from pyproject.toml
+# It replaces the version number on the line with the format `version = "1.2.3"`
 def set_current_package_version(version: str) -> None:
-    with open(VERSION_FILE_PATH, 'r+') as version_file:
-        updated_version_file_lines = []
+    with open(PYPROJECT_TOML_FILE_PATH, 'r+', encoding='utf-8') as pyproject_toml_file:
+        updated_pyproject_toml_file_lines = []
         version_string_found = False
-        for line in version_file:
-            if line.startswith('__version__'):
+        for line in pyproject_toml_file:
+            if line.startswith('version = '):
                 version_string_found = True
-                line = f"__version__ = '{version}'"
-            updated_version_file_lines.append(line)
+                line = f'version = "{version}"'
+            updated_pyproject_toml_file_lines.append(line)
 
         if not version_string_found:
             raise RuntimeError('Unable to find version string.')
 
-        version_file.seek(0)
-        version_file.write('\n'.join(updated_version_file_lines))
-        version_file.truncate()
+        pyproject_toml_file.seek(0)
+        pyproject_toml_file.write('\n'.join(updated_pyproject_toml_file_lines))
+        pyproject_toml_file.truncate()